RAG_system/requirements.txt
# Core Dependencies - Using BGE for better table retrieval
sentence-transformers>=2.3.1

# LLM
llama-index-llms-groq>=0.5.0

# Vector Store
chromadb>=1.5.0

# PDF Processing
pdfplumber==0.11.0

# UI
streamlit==1.31.0

# Utilities
python-dotenv==1.0.1
